Add `/data` and DataLayer constants
This commit introduces the data strategy that FlatJobs uses, whereby:
- raw data is saved in the `bronze` data layer
- parsed data is saved in the `silver` data layer
- consolidated data is saved in the `gold` data layer

This is using a medallion data lake architecture[^1], with all the pros and cons
that come with it.

Isn't this overkill for FlatJobs? Perhaps :) I'm choosing to save data
at the various stages for reproducibility when things invariably go
wrong - I can then replay with the same data. And looking further into
the future, by saving the raw data I'm giving myself the option of
extracting out more attributes that may be of interest.

[^1]:https://dataengineering.wiki/Concepts/Medallion+Architecture