Add `FlatJobs::Company`
This commit adds the `FlatJobs::Company` _abstract_ class, which is
responsible for coordinating the common approach for updating a
company's flat jobs data.

Each concrete company class is expected to inherit from
`FlatJobs::Company` and implement `#fetch_data`, `#data_file_type`, and
`#parse_jobs`.

The common update approach is to:
- fetch the raw company data
- persist the raw company data to `/data/bronze`
- parse the raw company data to flat jobs
- persist the flat jobs to `/data/silver`
- return success with the flat jobs

And if anything goes wrong, to return failure with the error message.

This class also adds the escape hatch to disable a company from being
processed if with a `<COMPANY_NAME>_DISABLED` ENV var being present.
This affords toggling this behaviour without needing to do commits/code
changes.